Apply a minimum padding to mapview overlays
This will not compound with safeAreaInsets, but ensure that devices
which don't have safeAreaInsets (like the iPhoneSE 3) will have some
padding between the overlay and the screen edge.

Now that the padding is applied to the Overlay view, some of the padding
previously added to views within the overlay view is now redundant.

remove double padding which is now redundant with safeAreaInsets