One camera up in the AppEnvironment. (#529)
With this change, if the iphone simulator is launched *and* the CarPlay window is already open, then starting navigation on the phone will also show navigation on the CarPlay window.
If the CarPlay window is *not* open on launch of the application in the iPhone simulator, it will not work.
As discussed with @Archdoog this is just to move things towards a good state. Adding the options to have a separate camera for each MapView will come along after launch of CarPlay can be done after CarPlay Navigation works from the get-go

Note: This is using the `SharedMapCamera` in `AppEnvironment`. However just having `@Published var camera : MapViewCamera = ...` will work too. This diff preserves `SharedMapCamera`, but technically speaking it is no longer necessary for the DemoApp.