Allow for optional accuracy in Speed
At the moment whole Speed entity is discarded if speed is valid but
accuracy is not. This case is very common with Xcode location
simulation, making it hard to use UserLocation from FerrostarCore as
it strips speed data out.

My suggestion is to pass speed through even if accuracy is nil.